+++ packages/glyph3d-core/src/picking/PickingSystem.js
@@ -232,8 +232,15 @@ export class PickingSystem {
         // instance index == arena slot).
         let byteSlots = null;
         if (byteMode) {
-            const placeholder = new StorageInstancedBufferAttribute(new Float32Array(4), 1);
-            byteSlots = registerByteSlotsNode(storage(placeholder, 'float', 1).toReadOnly().onObjectUpdate(({ object }, self) =>
+            // 'uint', matching the slot buffer's own type (glyphPipelineKernels: the
+            // slots array is instancedArray(..., 'uint')). This declaration is what the
+            // shader reads the memory AS — WGSL will bind the same bytes as array<f32>
+            // with no validation error, so a stale 'float' here silently reinterprets
+            // every count lane as a denormal (row 5 -> 0x5 -> ~7e-45 -> 0) and makes
+            // glyphVertex's .toFloat() a no-op. Near text is unaffected; the far-LOD UV
+            // collapses to texel (0,0) for every glyph. Nothing errors.
+            const placeholder = new StorageInstancedBufferAttribute(new Uint32Array(4), 1);
+            byteSlots = registerByteSlotsNode(storage(placeholder, 'uint', 1).toReadOnly().onObjectUpdate(({ object }, self) =>
                 (object && object.userData.glyphField && object.userData.glyphField._byteSlots) || self.value));
         }
